Centralize per-command middleware normalization
Move per-command middleware map normalization into a shared helper used by pipeline planning and pipeline map inspection.

This keeps command FQN normalization and empty middleware filtering in one place while leaving each caller's planning behavior unchanged.

Tests: dotnet test tests\TinyDispatcher.UnitTests\TinyDispatcher.UnitTests.csproj
